Add files via upload
https://www.kaggle.com/datasets/fedesoriano/body-fat-prediction-dataset
Context
Lists estimates of the percentage of body fat determined by underwater
weighing and various body circumference measurements for 252 men.

Educational use of the dataset
This data set can be used to illustrate multiple regression techniques. Accurate measurement of body fat is inconvenient/costly and it is desirable to have easy methods of estimating body fat that are not inconvenient/costly.